requirements.txt
# NHS Job Sponsorship Tracker — Python dependencies
# Local dev:      pip install -r requirements.txt
# Lambda package: deploy.sh handles this automatically

# ── Scraping ───────────────────────────────────────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# ── CV generation ──────────────────────────────────────────────────────────────
python-docx==1.1.0

# ── AWS SDK ────────────────────────────────────────────────────────────────────
boto3==1.34.0
botocore==1.34.0

# ── AI CV tailoring (optional — only needed if CV_MODE=ai) ────────────────────
anthropic==0.25.0
